Performance Improvement: Move Post Processing Outside of Each Track Insert (#37)
The post processing synchronizer takes quite a long time compared to the other synchronizers. It technically only needs to be run once at the end of processing. I ran it after each insert so that if a user did pause the processing early, the data from the post processor would exist. However, most users likely won't do this and it adds a considerable performance hit to ingesting tracks (which is arguably a worse user experience than missing data until fully processed).

Previously, it took around 30+ minutes to ingest 2000+ tracks
Now, it takes around 10 minutes

The other synchronizers/generators appear to be operating as expected, and there is not much performance improvements we can make to them because of the limitations of the database/EF queries. One option is to pipeline the synchronizers (they take a bit longer than the generators), but that would likely yield little gain and require considerable investigation into whether or not it is safe (from a data integrity / query blocking perspective) to actually pipeline the dataflow.