Fix Repeat Ingestions of Old Tracks (#52)
Seems to be something to do with the update dates.

- [x] But first, add a readme for the configs you added in the previous
PR

When the last modified date gets out of sync (in this case I believe I
modified a file, then modified it back to its original state), we hit a
new "None" branch in the code that runs slightly slower than the
"Skipped" branch. The reason these files repeatedly ingest is because we
don't update the modified date, so once we hit the new "None" branch,
those files will continue to hit that branch until the modified date
changes manually.