Remove Automatic Migration System (#53)
Not super useful, if you need to update your database schema, you can
just execute it yourself. We also don't have a drop-re-initialize system
in place if the auto-migrate fails, so we are already relying on the
user to run sql manually for their chosen database provider. We may
re-visit this in the future.

---------

Co-authored-by: GitHub Actions <actions@github.com>